Improve DecodeMSADPCMBlock performance
A few performance fixes for MSADPCM decoding. In my testing,
this was able to reduce MSADPCM decode times to around 1/3rd of
what they were previously. (Naddachance startup from 18s to 6s).

Note that the force inline stuff is gcc specific. It should probably be
wrapped up in preprocessor macros so that things will still compile on
MSVC.